Editor: Move PluginPostPublishPanel and PluginPrePublishPanel to editor package
